refactor: rename schema targets, pass api urls as env variables
