+++ .env.template
@@ -1,4 +1,6 @@
 GITHUB_AUTH= # for lerna changelog
 AUTH_TOKEN= # for github labels
+MC_API_URL=https://mc-api.commercetools.com/graphql
+MC_PROXY_URL=https://mc.commercetools.com/api/graphql
 MC_ACCESS_TOKEN= # for schema introspection
 CTP_PROJECT_KEY= # one of your CTP projects where you have access to
