+++ packages/application-shell/src/components/authenticated/authenticated.mc.graphql
