+++ packages/application-shell/src/components/fetch-project/fetch-project.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { GRAPHQL_TARGETS } from '@commercetools-frontend/constants';
 import { renderApp, waitForElement } from '../../test-utils';
-import ProjectQuery from './fetch-project.users.graphql';
+import ProjectQuery from './fetch-project.mc.graphql';
 import FetchProject, {
   mapAllAppliedToObjectShape,
   mapAllAppliedToGroupedObjectShape,
