+++ packages/application-shell/src/components/fetch-user/fetch-user.mc.graphql
@@ -1,4 +1,4 @@
-#import "../project-switcher/project-switcher-fragments.users.graphql"
+#import "../project-switcher/project-switcher-fragments.mc.graphql"
 
 query LoggedInUserQuery {
   user: me {
