+++ packages/application-shell/src/components/project-switcher/project-switcher-fragments.mc.graphql
