+++ packages/application-shell/src/hooks/use-applications-menu/fetch-applications-menu.proxy.graphql
