+++ packages/application-shell/src/hooks/use-applications-menu/use-applications-menu.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import upperFirst from 'lodash/upperFirst';
 import { reportErrorToSentry } from '@commercetools-frontend/sentry';
 import { renderApp, waitForElement, wait } from '../../test-utils';
-import FetchApplicationsMenu from './fetch-applications-menu.menus.graphql';
+import FetchApplicationsMenu from './fetch-applications-menu.proxy.graphql';
 import useApplicationsMenu from './use-applications-menu';
 
 jest.mock('@commercetools-frontend/sentry');
