+++ packages/application-shell/src/utils/select-user-id/select-user-id.mc.graphql
