+++ schemas/mc.json
@@ -7799,7 +7799,7 @@
   "extensions": {
     "graphql-config": {
       "source": "https://mc-api.commercetools.com/graphql",
-      "timestamp": "Tue Sep 10 2019 18:19:08 GMT+0200 (Central European Summer Time)"
+      "timestamp": "Wed Sep 11 2019 09:40:28 GMT+0200 (Central European Summer Time)"
     }
   }
 }
