+++ schemas/proxy.json
@@ -1498,7 +1498,7 @@
   "extensions": {
     "graphql-config": {
       "source": "https://mc.commercetools.com/api/graphql",
-      "timestamp": "Tue Sep 10 2019 18:19:09 GMT+0200 (Central European Summer Time)"
+      "timestamp": "Wed Sep 11 2019 09:40:29 GMT+0200 (Central European Summer Time)"
     }
   }
 }
